requirements/base.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# API Documentation (Swagger)
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Async Support
uvicorn>=0.27.0
gunicorn>=21.2.0

# Celery
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
flower>=2.0.1

# Redis
redis>=5.0.1
django-redis>=5.4.0

# RabbitMQ
pika>=1.3.2
kombu>=5.3.4

# AWS S3
boto3>=1.34.0
django-storages>=1.14.2

# AI Services
openai>=1.10.0
google-generativeai>=0.3.2

# Machine Learning
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Authentication
djangorestframework-simplejwt>=5.3.1
PyJWT>=2.8.0

# Validation & Serialization
pydantic>=2.5.3
python-dateutil>=2.8.2

# Environment
python-dotenv>=1.0.0
python-decouple>=3.8

# Monitoring
django-prometheus>=2.3.1
sentry-sdk>=1.39.1

# Utils
Pillow>=10.2.0
python-slugify>=8.0.1
uuid>=1.30

# Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
selenium>=4.15.0
webdriver-manager>=4.0.1
